refactor(ui-design-tool): split `UIDesignTool` class
