+++ packages/ui-design-tool/components/UIDesignToolProvider/UIDesignToolProvider.context.ts
@@ -6,7 +6,7 @@ const [UIDesignToolContextProvider, ...hooks] = constate(
   (value) => value.controllerInterface,
   (value) => value.dataInterface,
   (value) => value.selectorInterface,
-  (value) => value.subscriptionInterface,
+  (value) => value.subscriberInterface,
 
   (value) => value.dispatcher,
   (value) => value.instanceId,
@@ -31,12 +31,18 @@ export const [
   // 패키지 외부 제공 (재조정 범위에 따라 내부에서도 사용)
   // - 은닉한 프로퍼티는 노출해서는 안됨
   // - 재조정 범위는 줄이되 패키지 사용자를 고려해 너무 많은 hook이 생기지 않도록 함
+
+  /** @see ModelStore */
   useUIController,
+  /** @see ModelStore */
   useUIData,
+  /** @see ElementSelector */
   useUISelector,
-  useUISubscription,
+  /** @see DataSubscriber */
+  useUISubscriber,
 
   // 패키지 내부 사용
+
   useDispatcher,
   useInstanceID,
   useBrowserMeta,
