+++ packages/ui-design-tool/hooks/useUISubscriber.tsx
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-restricted-imports
+export { useUISubscriber as default } from '@/components/UIDesignToolProvider/UIDesignToolProvider.context';
