--- packages/ui-design-tool/hooks/useUISubscription.tsx
@@ -1,2 +1,0 @@
-// eslint-disable-next-line no-restricted-imports
-export { useUISubscription as default } from '@/components/UIDesignToolProvider/UIDesignToolProvider.context';
